feat: add dry-run option (#78)

+++ hashin.py
@@ -15,6 +15,7 @@ from itertools import chain
 
 import pip_api
 from packaging.version import parse
+import difflib
 
 if sys.version_info >= (3,):
     from urllib.request import urlopen
@@ -75,6 +76,12 @@ parser.add_argument(
     action='store_true',
     default=False,
 )
+parser.add_argument(
+    '--dry-run',
+    help="Don't touch requirements.txt and just show the diff",
+    action='store_true',
+    default=False,
+)
 
 
 major_pip_version = int(pip_api.version().split('.')[0])
@@ -101,7 +108,7 @@ def _download(url, binary=False):
     # Note that urlopen will, by default, follow redirects.
     status_code = r.getcode()
 
-    if status_code >= 301 and status_code < 400:
+    if 301 <= status_code < 400:
         location, _ = cgi.parse_header(r.headers.get('location', ''))
         if not location:
             raise PackageError("No 'Location' header on {0} ({1})".format(
@@ -137,6 +144,7 @@ def run_single_package(
     python_versions=None,
     verbose=False,
     include_prereleases=False,
+    dry_run=False,
 ):
     restriction = None
     if ';' in spec:
@@ -159,7 +167,6 @@ def run_single_package(
     package = data['package']
 
     maybe_restriction = '' if not restriction else '; {0}'.format(restriction)
-    new_lines = ''
     new_lines = '{0}=={1}{2} \\\n'.format(
         package,
         data['version'],
@@ -175,17 +182,31 @@ def run_single_package(
             new_lines += ' \\'
         new_lines += '\n'
 
-    if verbose:
-        _verbose('Editing', file)
     with open(file) as f:
-        requirements = f.read()
+        old_requirements = f.read()
     requirements = amend_requirements_content(
-        requirements,
+        old_requirements,
         package,
         new_lines
     )
-    with open(file, 'w') as f:
-        f.write(requirements)
+    if dry_run:
+        if verbose:
+            _verbose('Dry run, not editing ', file)
+        print(
+            "".join(
+                difflib.unified_diff(
+                    old_requirements.splitlines(True),
+                    requirements.splitlines(True),
+                    fromfile="Old",
+                    tofile="New",
+                )
+            )
+        )
+    else:
+        with open(file, 'w') as f:
+            f.write(requirements)
+        if verbose:
+            _verbose('Editing', file)
 
 
 def amend_requirements_content(requirements, package, new_lines):
@@ -498,6 +519,7 @@ def main():
             args.python_version,
             verbose=args.verbose,
             include_prereleases=args.include_prereleases,
+            dry_run=args.dry_run,
         )
     except PackageError as exception:
         print(str(exception), file=sys.stderr)
